move files in packages
