--- app/src/main/java/com/example/moonlight/CartViewModel.kt
@@ -1,7 +1,0 @@
-package com.example.moonlight
-
-import androidx.lifecycle.ViewModel
-
-class CartViewModel : ViewModel() {
-    // TODO: Implement the ViewModel
-}
