+++ app/src/main/java/com/example/moonlight/ui/account/AccountViewModel.kt
@@ -1,4 +1,4 @@
-package com.example.moonlight
+package com.example.moonlight.ui.account
 
 import androidx.lifecycle.ViewModel
