+++ app/src/main/java/com/example/moonlight/ui/categoria/CategoriaFragment.kt
@@ -1,4 +1,4 @@
-package com.example.moonlight
+package com.example.moonlight.ui.categoria
 
 import androidx.lifecycle.ViewModelProvider
 import android.os.Bundle
@@ -6,25 +6,26 @@ import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import com.example.moonlight.R
 
-class AccountFragment : Fragment() {
+class CategoriaFragment : Fragment() {
 
     companion object {
-        fun newInstance() = AccountFragment()
+        fun newInstance() = CategoriaFragment()
     }
 
-    private lateinit var viewModel: AccountViewModel
+    private lateinit var viewModel: CategoriaViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?,
     ): View? {
-        return inflater.inflate(R.layout.fragment_account, container, false)
+        return inflater.inflate(R.layout.fragment_categoria, container, false)
     }
 
     override fun onActivityCreated(savedInstanceState: Bundle?) {
         super.onActivityCreated(savedInstanceState)
-        viewModel = ViewModelProvider(this).get(AccountViewModel::class.java)
+        viewModel = ViewModelProvider(this).get(CategoriaViewModel::class.java)
         // TODO: Use the ViewModel
     }
