+++ app/src/main/java/com/example/moonlight/ui/main/MainViewModel.kt
@@ -1,0 +1,12 @@
+package com.example.moonlight.ui.main
+
+import androidx.lifecycle.ViewModel
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.asStateFlow
+
+class MainViewModel : ViewModel() {
+    private val _uiState = MutableStateFlow(true)
+    val uiState: StateFlow<Boolean>
+        get() = _uiState.asStateFlow()
+}
