+++ app/src/main/res/layout/fragment_cart.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".CartFragment">
+    tools:context=".ui.cart.CartFragment">
 
     <TextView
         android:layout_width="match_parent"
